no message
git-svn-id: https://svn.code.sf.net/p/khc/svn/branches/unicode@2249 ab1598f5-0ef3-491f-8894-8c1f2390833e

+++ core/kh_lib/kh_morpho/win32/mecab.pm
@@ -90,6 +90,7 @@ sub _mecab_run{
 	$self->_mecab_store($t) if length($t);
 	$self->_mecab_store_out;
 
+	return 1 unless -e $self->{target_temp};
 	return 1 unless -s $self->{target_temp} > 0;
 	unlink $self->{output_temp} if -e $self->{output_temp};
